replace links to clojure core api to source code in local
For instance, we replace
  [[https://clojure.github.io/clojure/clojure.core-api.html#clojure.core/dosync][dosync]]
by
  [[clojure:core.clj::(defmacro dosync][dosync]]

where [[clojure]] is a org abbreviation link declared a the top of the
org file with the line:

  #+LINK: clojure ./clojure/src/clj/clojure/

Now when we follow a link to a core function (for instance), we jump
into Clojure source code locally.  To make it works you have to clone
Clojure in your computer.