update line wrapping in paragraphs containing links
we've used `pandoc` with the default line wrapping at 72 characters
per lines.  This gives the result expected everywhere but on the
paragraphs containing links.  Indeed, when using org-mode (inside
emacs) the url part of the links is made invisible and org-mode offer
another (hard) line wrapping that take this invisibility into account
and just count the visible characters to do the (hard) line wrapping.